pyproject.toml
[tool.poetry]
name = "file-update-discord"
version = "0.1.0"
description = ""
authors = ["Kyuzial <kyuzial@protonmail.com>"]
readme = "README.md"
packages = [
    { include = "file_update_discord", from = "src"},
]

[tool.poetry.dependencies]
python = "^3.11"
discord-py = "^2.3.1"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.0.280"
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["src"]
